Migrating to Swift 3.1.

+++ runtime/Swift/Sources/Antlr4/atn/ProfilingATNSimulator.swift
@@ -57,7 +57,7 @@ public class ProfilingATNSimulator: ParserATNSimulator {
         decisions[decision].timeInPrediction += (stop - start)
         decisions[decision].invocations += 1
 
-        var SLL_k: Int64 = _sllStopIndex - _startIndex + 1
+        var SLL_k: Int64 = Int64(_sllStopIndex - _startIndex + 1)
         decisions[decision].SLL_TotalLook += SLL_k
         decisions[decision].SLL_MinLook = decisions[decision].SLL_MinLook == 0 ? SLL_k : min(decisions[decision].SLL_MinLook, SLL_k)
         if SLL_k > decisions[decision].SLL_MaxLook {
@@ -67,7 +67,7 @@ public class ProfilingATNSimulator: ParserATNSimulator {
         }
 
         if _llStopIndex >= 0 {
-            var LL_k: Int64 = _llStopIndex - _startIndex + 1
+            var LL_k: Int64 = Int64(_llStopIndex - _startIndex + 1)
             decisions[decision].LL_TotalLook += LL_k
             decisions[decision].LL_MinLook = decisions[decision].LL_MinLook == 0 ? LL_k : min(decisions[decision].LL_MinLook, LL_k)
             if LL_k > decisions[decision].LL_MaxLook {
